refactor(httpwebrequest): extract shared external-segment helper
Pull the duplicated external-segment creation (3 sites) and DT-header
injection (2 sites) out of GetRequestStreamWrapper and GetResponseWrapper
into a shared internal HttpWebRequestExternalSegment helper. Behavior
unchanged; brace-style and a stray computed-but-unused local cleaned up.
